@@ -1,10 +1,10 @@
 name: flutter_radar_chart
 description: Basic radar chart for Flutter. This chart is intended to work out of the box, adjusting itself to the given data. The design of the chart was inspired by the Python Graph Gallery.
-version: 0.1.6
+version: 0.2.0
 homepage: https://github.com/danpanaite/flutter_radar_chart
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
